Adds AppVeyor integration
Adds AppVeyor integration, so each commit is built with Lua 5.1.5, 5.2.4 and 5.3.1
on 32 and 64 bits, using Visual Studio 2015, 2013 and 2012.

Visual Studio 2010 and 2008 are not directly supported because it seems the versions
available on AppVeyor do not support 64 bit builds.
Even though the 32 buit builds work fine, in order to simplify the matrix, we don't
build with those compilers for the time being.